Update README.md
# Overview 
This repository contains the data generated by the Deep Learning for Automated Monitoring of Crystallization Processes.

# Train/Validation/Test Data 
The base dataset is compiled from the 800 annotated images of mannitol and sucrose crystals, and their accompanying configuration file. This dataset was divided into training and testing subsets in a 9:1 ratio, resulting in 1,335 training instances.